All 100 codes are added!

+++ lib/BetterSunshineEngine/lib/sms_interface/include/JSystem/JDrama/JDRActor.hxx
@@ -24,7 +24,7 @@ namespace JDrama {
 
         TVec3f mScale;     // 24
         TVec3f mRotation;  // 30
-        u32 _3C;           // 3C
+        f32 _3C;           // 3C
         TNameRef *mRootRef;
     };
